backend/requirements.txt
numba
numpy==1.23.5
torch
tqdm
more-itertools
tiktoken==0.3.1
ffmpeg-python==0.2.0
websockets
pywhispercpp==1.0.7
openai==0.27.2
python-dotenv==1.0.0
#pytorch